pyproject.toml
[tool.poetry]
name = "pixeltable"
version = "0.1.2"
description = "Pixeltable: a table interface to image and video data"
authors = ["Marcel Kornacker <marcelk@gmail.com>"]
readme = "README.md"
exclude = [
    ".pytype",
    ".pytest_cache",
    "pixeltable/.pytest_cache",
    "pixeltable/tests/data"
]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.24.1"
pandas = "^1.5.3"
pillow = "^9.4.0"
opencv-python-headless = "^4.7.0.68"
tqdm = "^4.64.1"
jmespath = "^1.0.1"
regex = "^2022.10.31"
cloudpickle = "^2.2.1"
psycopg2-binary = "^2.9.5"
psutil = "^5.9.5"
sqlalchemy = {extras = ["mypy"], version = "^2.0.23"}
sqlalchemy-utils = "^0.41.1"
autonomi-nos = {version = "^0.0.9", optional = true}
pgvector = "^0.2.1"
av = ">=10.0.0"
pgserver = "0.0.4"
torch = {version = ">=2.0.0", optional = true}
torchvision = {version = ">=0.16.0", optional = true}
pyarrow = {version = ">=13.0.1", optional = true}
openai = {version = "^1.0.0", optional = true}
boto3 = {version = "^1.17", optional = true}

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
pytest-xdist = {extras = ["psutil"], version = "^3.3.1"}
torch = "^2.1.1"
torchvision = "^0.16.1"
pyarrow = ">=13.0.1"
pycocotools = "^2.0.7"
ipykernel = "^6.27.1"
nbmake = "^1.4.6"
mypy = "^1.7.1"
types-pillow = "^10.1.0.2"
mkdocs = "^1.5.3"
mkdocstrings = {extras = ["python"], version = "^0.24.0"}
mkdocs-material = "^9.5.3"
mkdocs-jupyter = "^0.24"
black = "^23.12.1"
bs4 = "^0.0.2"
requests = "^2.31.0"

[tool.poetry.extras]
s3 = ["boto3"]
openai = ["openai"]
nos = ["autonomi-nos"]
torch = ["torch", "torchvision", "pyarrow"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
